Title: 최소 에너지 비용, Time: 148ms, Memory: 12MB, Status: Passed - Codetree

+++ 251028/최소 에너지 비용/minimum-energy-cost.java
@@ -13,21 +13,21 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int n = Integer.parseInt(br.readLine());
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int[] arr = new int[n - 1];
+        long[] arr = new long[n - 1];
 
         for (int i = 0; i < n - 1; i++) {
-            arr[i] = Integer.parseInt(st.nextToken());
+            arr[i] = Long.parseLong(st.nextToken());
         }
 
 
-        int[] arr2 = new int[n];
+        long[] arr2 = new long[n];
 
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < n; i++) {
-            arr2[i] = Integer.parseInt(st.nextToken());
+            arr2[i] = Long.parseLong(st.nextToken());
         }
 
-        int[] arr3 = new int[n];
+        long[] arr3 = new long[n];
 
         arr3[0] = arr2[0];
         for (int i = 1; i < n - 1; i++) {
